Move interface to /beacon/p2p/
