+++ beacon-chain/p2p/peers/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
         "//beacon-chain/core/helpers:go_default_library",
         "//beacon-chain/p2p/peers/peerdata:go_default_library",
         "//beacon-chain/p2p/peers/scorers:go_default_library",
-        "//proto/beacon/p2p/interfaces:go_default_library",
+        "//proto/beacon/p2p:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//shared/params:go_default_library",
         "//shared/rand:go_default_library",
