+++ beacon-chain/p2p/peers/peerdata/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers/peerdata",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
-        "//proto/beacon/p2p/interfaces:go_default_library",
+        "//proto/beacon/p2p:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/beacon/rpc/v1:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
