+++ beacon-chain/p2p/testing/BUILD.bazel
@@ -21,7 +21,7 @@ go_library(
         "//beacon-chain/p2p/encoder:go_default_library",
         "//beacon-chain/p2p/peers:go_default_library",
         "//beacon-chain/p2p/peers/scorers:go_default_library",
-        "//proto/beacon/p2p/interfaces:go_default_library",
+        "//proto/beacon/p2p:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
         "@com_github_ethereum_go_ethereum//crypto:go_default_library",
