+++ beacon-chain/p2p/testing/p2p.go
@@ -23,7 +23,7 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/p2p/encoder"
 	"github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers"
 	"github.com/prysmaticlabs/prysm/beacon-chain/p2p/peers/scorers"
-	p2pInterfaces "github.com/prysmaticlabs/prysm/proto/beacon/p2p/interfaces"
+	"github.com/prysmaticlabs/prysm/proto/beacon/p2p"
 	ethpb "github.com/prysmaticlabs/prysm/proto/eth/v1alpha1"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/proto"
@@ -39,7 +39,7 @@ type TestP2P struct {
 	DelaySend       bool
 	Digest          [4]byte
 	peers           *peers.Status
-	LocalMetadata   p2pInterfaces.Metadata
+	LocalMetadata   p2p.Metadata
 }
 
 // NewTestP2P initializes a new p2p test service.
@@ -338,7 +338,7 @@ func (p *TestP2P) ForkDigest() ([4]byte, error) {
 }
 
 // Metadata mocks the peer's metadata.
-func (p *TestP2P) Metadata() p2pInterfaces.Metadata {
+func (p *TestP2P) Metadata() p2p.Metadata {
 	return p.LocalMetadata.Copy()
 }
