+++ beacon-chain/sync/BUILD.bazel
@@ -65,7 +65,7 @@ go_library(
         "//beacon-chain/state/interface:go_default_library",
         "//beacon-chain/state/stategen:go_default_library",
         "//cmd/beacon-chain/flags:go_default_library",
-        "//proto/beacon/p2p/interfaces:go_default_library",
+        "//proto/beacon/p2p:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "//proto/beacon/p2p/v1/wrapper:go_default_library",
         "//proto/eth/v1alpha1:go_default_library",
