+++ proto/beacon/p2p/v1/wrapper/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1/wrapper",
     visibility = ["//visibility:public"],
     deps = [
-        "//proto/beacon/p2p/interfaces:go_default_library",
+        "//proto/beacon/p2p:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@org_golang_google_protobuf//proto:go_default_library",
