+++ proto/beacon/p2p/v1/wrapper/metadata.go
@@ -2,7 +2,7 @@ package wrapper
 
 import (
 	"github.com/prysmaticlabs/go-bitfield"
-	"github.com/prysmaticlabs/prysm/proto/beacon/p2p/interfaces"
+	"github.com/prysmaticlabs/prysm/proto/beacon/p2p"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
 	"google.golang.org/protobuf/proto"
 )
@@ -38,7 +38,7 @@ func (m MetadataV0) IsNil() bool {
 }
 
 // Copy performs a full copy of the underlying metadata object.
-func (m MetadataV0) Copy() interfaces.Metadata {
+func (m MetadataV0) Copy() p2p.Metadata {
 	return WrappedMetadataV0(proto.Clone(m.md).(*pb.MetaDataV0))
 }
 
@@ -85,7 +85,7 @@ func (m MetadataV1) IsNil() bool {
 }
 
 // Copy performs a full copy of the underlying metadata object.
-func (m MetadataV1) Copy() interfaces.Metadata {
+func (m MetadataV1) Copy() p2p.Metadata {
 	return WrappedMetadataV1(proto.Clone(m.md).(*pb.MetaDataV1))
 }
